Autogenerate Attach/Launch Dialogs [2/N]: OCaml Debugger
Summary:
Replaces the `OCamlLaunchUIComponent` and `OCamlLaunchProvider` with the `AutoGenLaunchAttachProvider` and `AutoGenLaunchAttachProvider`.

Now it looks like this:
https://pxl.cl/cqLD
Before it looked like this:
https://pxl.cl/cqLx

The notable changes are that certain properties say (required) next the name if they are indeed required. Additionally, the Launch button is disabled until the user has provided the required fields.

Reviewed By: pittsw

Differential Revision: D7407806

fbshipit-source-id: 9a372bcb20eb86f38b9b5dd0782ede720b4a170a