Debugger: Get rid of evaluation expression provider APIs
Summary:
That was never useful actually & the default implementation is used for all the integrations that work today.

Also, fix an issue with the default evaluation expression provider.

Hence, removing this APIs cleans the code, results in faster results (no re-routing to language APIs) + avoids depending on LSP/languages for finding debug expression.

Reviewed By: ebluestein

Differential Revision: D7460371

fbshipit-source-id: b21f9a20ce022fa491e14af00875e513aacde79b