Add onToggleCoverage and DynamicView to LspLanguageService
Summary:
This hooks up LspLanguageService to dynamic view, and defines a new Lsp capability method "workspace/DynamicView", which sets a flag on the hack server.

The server then adds this capability in the ocaml diffs.

This has a weird consequence: typeCoverage itself only updates on save, but errors update instantly, which leads to some weird underlines. I'll probably address this in another diff above.

Reviewed By: arxanas

Differential Revision: D7410457

fbshipit-source-id: ffc3153fba98a13e65bcd3fb2989539ad0db4474