Add a command to toggle focus between terminal and editor.
Summary:
When using an external Terminal application, I usually `cmd-tab` out of Nuclide
to my Terminal and then go back. In introducing this command, I would like to
create the same experience using the built-in Terminal with less of a context
switch.

Because I had to pick something, I established an initial keybinding of
`ctrl-!`, but I'm happy to change it if someone has a more intuitive (and
available) suggestion.

Reviewed By: ebluestein

Differential Revision: D7463169

fbshipit-source-id: 0f557c3266f33d52e92e2f39b8ac050de097efd6