Handle null filenames from fs.watch
Summary:
`fs.watch` may sometimes emit `null` as a changed filename (unclear when, but it's explicitly mentioned as a caveat in https://nodejs.org/api/fs.html#fs_filename_argument)

For filename watches we can proceed as usual, but for directories we have to be defensive and just emit nothing.

Flow doesn't know this unfortunately. This is the only callsite that's affected.

Reviewed By: stepanhruda

Differential Revision: D7480057

fbshipit-source-id: 8f444e2c5f57662c3785c385f980c08ecf76f18f