fix homebrew openssl path
Summary:
Not sure what version of homebrew placed openssl in /opt/homebrew/bin, but the current installer places it in /usr/local/opt/openssl/bin.

I'm assuming openssl has not been linked in homebrew (made the default), or the path addition would be unnecessary.

I was unable to run nuclide-start-server on High Sierra (10.13.4) until I made this change. This seems to be because LibreSSL does not support environment variables in the included openssl.cnf.
Closes https://github.com/facebook/nuclide/pull/1492

Differential Revision: D7492257

Pulled By: hansonw

fbshipit-source-id: 112eb0a6050d4dcf3d9c1cb31f419c618668012a