don't open on start for deep links
Summary: this looked super clowny when opening on demand and the connect menu briefly flashes open just to be closed. we needed to check if we were opening from a deep link *before* we open the connect menu in the breadcrumbs activation, so we couldn't use the deep link service. so this is just a flag set in the deep link utils.

Reviewed By: hansonw

Differential Revision: D7481041

fbshipit-source-id: 32f91c14f43d2b57b457d3251a6a3ee29a8f5185