Fix issue with QueuedAckTransport on Windows + Atom 1.25
Summary:
pelmers (and mostafaeweda / hchau-fb) helped track down an extremely weird regression for Windows users on Atom 1.25 where sometimes messages would get stuck in the event queue :O

The workaround is to call `process.activateUvLoop()`, Electron's escape hatch for force triggering the Node event loop. If you look in the Electron code there are actually several global functions that are wrapped to do this - so the Node integration seems like it's not as robust as we would imagine.

https://github.com/electron/electron/blob/7ef69a5af595c9c0eef27bdf9446103776917aec/lib/common/init.js#L29

In this particular case `ACK` messages are scheduled using `setTimeout` (which runs on the Chrome event loop) but then the actual message sending happens on the Node event loop. Also the `setTimeout` is triggered via the Node event loop so perhaps going back and forth triggers some weird edge case with the TLS/SSL code? I honestly don't know.

Reviewed By: semmypurewal

Differential Revision: D7508287

fbshipit-source-id: 9e59b88ddffef2c629bf6f2911f322908cb665fd