Fix stderr piping
Summary:
In D7474677 I made the mistake of piping the child process's stderr without disconnecting it, which causes big-dig to hang during normal startup. I don't think there's a great way of doing this safely actually, since the child process may run into errors trying to write to a detached pipe.

The safest alternative is to direct stderr into a tempfile and then read it back.

Reviewed By: wanderley

Differential Revision: D7513268

fbshipit-source-id: 6d15231b4b661942140acc4696fdde6949986c55