Add descriptionMarkdown field in autocomplete suggestions
Summary:
A recent version of `autocomplete-plus` added support for `descriptionMarkdown` in suggestions, which does what it says on the label. This adds it to the Flow types and RPC def so that providers can start using it if they wish to.

I wouldn't recommend using this for the LSP converter without some more testing (or maybe some custom CSS will do the job). It seems to jump around a lot.

Reviewed By: ljw1004

Differential Revision: D7538818

fbshipit-source-id: ce53f1aefbf10a298eca7a8bcceaeafb73b5de09