Remove scroll preservation feature
Summary:
While nice to have, mostafaeweda already demonstrated that this was a clear CPU hog in D6562222.
(Even if we don't update the navigation stack on scroll, just measuring the scroll position requires a DOM measurement).

While investigating T25790515, I'd like to clean up the behavior of the navigation stack to be more consistent.

Having to keep the scroll position around complicates things significantly so I'd like to just remove it - to be honest I don't think anyone will notice.

Reviewed By: hchau-fb

Differential Revision: D7495651

fbshipit-source-id: 69d6b77f095f48664340087e976ec4486cf2f0bd