Try to reduce memory usage of the workers
Summary:
The worker process often hangs on to 1GB+ of memory after initial indexing is complete, and I couldn't figure out why! I did some memory profiling and it appears there's a bunch of stuff in Rx that hangs on to the file index + export cache... I'll have to figure out why Rx is doing that but the easiest workaround is to null out the underlying cache values manually.

Also, handle cases where the worker dies (previously we would just keep trying to send the dead worker messages - oops..)

Reviewed By: wanderley

Differential Revision: D7548971

fbshipit-source-id: 07ba8531d2e04bd15e422f5da1694ca8d00352d0