Fix flakiness in fuzzy file search test
Summary:
I added a feature a while back to include the directory name when searching for files (so queries like `www/asdfasdf` work as expected).
Unexpectedly, this causes test flakiness because fixture directories have a random directory name (which may partially match the test queries!)
To remedy this we'll force a consistent directory name by adding a "project" directory to the fixture.

Reviewed By: pelmers

Differential Revision: D7571022

fbshipit-source-id: d4b3d6b25d56d468cc44c3ba9606f2d03b816c76