Add an example for HighlightedCode (using Async React)
Summary:
This adds an example to the sample UI playground (with async rendering enabled)! The API is marked as `unstable` but it seems to work with 16.2, even.

The API for async React is surprisingly simple - what this means is that once the button to add code is pressed, the resulting re-render can be scheduled by React to happen over multiple frames. You don't see the updated DOM until all the updates have finished.

Because `HighlightedCode` tokenizes lines individually, this means that this effectively does what TextEditors do (highlight lines in batch) but natively with React!

(Note: in 16.3 `React.unstable_AsyncComponent` becomes `React.unstable_AsyncMode`.

Reviewed By: matthewwithanm

Differential Revision: D7540677

fbshipit-source-id: 8880165a7f606693a6af5bb3c298e93d11f57f36