Switch Tunnel usages to ResolvedTunnel
Summary:
This is an object with the host URI normalized, we only want to use `Tunnel` at the client boundaries of the package from now on.

You won't be able to see descriptions in the UI because `ResolvedTunnel` doesn't have it (it will be based on subscriptions). It will start working 5 commits later, I promise.

Reviewed By: aadsm

Differential Revision: D7436791

fbshipit-source-id: 007ab73d731f65626b01ebe532ad7740d3a7c2d4