Switch to new nuclide-ssh-tunnel API
Summary:
More code deletion. We can now have multiple call sites request tunneling and turn it off when no one requires it anymore. But it's not that important since we will simply want to automatically turned on.

While working on this, I realized we could probably implement the core of nuclide-ssh-tunnel this way instead of redux and the logic would end up much simpler. But I didn't know what I know now when I started on this stack.

Reviewed By: aadsm

Differential Revision: D7436842

fbshipit-source-id: dba7bd188c2958f5e972d2af7587c59b709c55ca