Use Node v9 for ESLint on CircleCI
Summary:
It seems that the Prettier ESLint plugin doesn't work well on Node v6/v7 and occasionally crashes and/or OOMs. We're seeing this quite frequently on CircleCI now.

see: https://github.com/prettier/prettier/issues/3457

It appears that the best workaround is to use a higher Node version when running ESLint. I was able to confirm on a Linux machine that this uses substantially less memory.
Closes https://github.com/facebook/nuclide/pull/1499

Reviewed By: captbaritone

Differential Revision: D7587363

Pulled By: hansonw

fbshipit-source-id: 86d245572952ec881e9f185ae0a56e151d489ae0