Show successor info in smartlog
Summary:
A revision can be succeeded by a different one in a variety of conditions if you aren't careful to restack children. Interactive Smartlog should usually restack, but it's useful for us to show this information when we have it. This also lets us say e.g. "Landed as f12ab42" for a commit who is succeeded by a public commit.

We actually already had this information fetched every time, I'm not sure why we weren't using it.

Note that this doesn't work for rebases in progress or when in merge conflicts. That's actually the case I case most about, so I'm probably going to see if we can hack something for that case. The problem is that successor info isn't given until the operation is complete, so if you're in a merge conflict it won't report existing successors from previously resolved conflicts in the stack (for example).

Reviewed By: tjfryan

Differential Revision: D7504115

fbshipit-source-id: d299e6ef0dd7b8be2088a339a85a81cfbc3e9ca8