Add link to successor in commit info view
Summary:
If you're looking at a commit in the commit info view which has a successor, we can display that and let you click there to open that commit in the commit info view.

Note that some commits may not be in smartlog, so opening the newer version may not work. This mostly happens with commits that are now landed.

Reviewed By: tjfryan

Differential Revision: D7506742

fbshipit-source-id: 540426f56b10502e18306f8877ebc7a94853f074