Resolve the realpath of file args
Summary:
People often do `atom ~/folder/file` which may be a symlink; this doesn't play well with Nuclide, which expects all remote paths to be realpaths.

Sometimes the new path may not exist yet - in which case we need to try to resolve the realpath of all the possible prefixes.

Reviewed By: bolinfest

Differential Revision: D7610207

fbshipit-source-id: 1fdcd394fb9fc288389c210dad133096d04e4926