Introduce nuclide-connections executable to list connections to Nuclide server.
Summary:
It is useful to have a utility to check the current connections to the Nuclide
server. This does the most basic thing, which is enumerates the remote root
folders in Atom clients for the Nuclide server.

Note that this follows the general pattern of how the remote `atom` command
is implemented. It takes a lot of inspiration from
`nuclide/pkg/nuclide-remote-atom-rpc/bin/main.js`.

Note that `CommandClient.js` assumed that if a connection could not be made
to the local Nuclide server, then `process.exit()` should be called. I had to
refactor some code to make that behavior optional because
`nuclide-connections` should still return with a zero exit code in that case (and
without spamming stderr).

Reviewed By: hansonw

Differential Revision: D7598900

fbshipit-source-id: d266f2752d7f59bc1822495fa94fd2051c043a47