Ability to specify which node version to use when running jest tests
Summary:
Nuclide overwrites PATH and makes it use an older version of node that is not compatible with the debugger. Also it's generally better to have control over which node version we run.
This diff adds an optional config entry that specifies where to find the node binary that'll be used to run all jest tests

Reviewed By: captbaritone

Differential Revision: D7631298

fbshipit-source-id: 59ccfc5149a7710bdd8984257319947e7d8589fe