Fix commit change watcher in HgService so it works with the share extension.
Summary:
Nuclide's `HgService` monitors `00changelog.i`, `obsstore`, and `inhibit` in
`.hg/store` for changes as a heuristic to infer commit changes in an Hg repo.
Previously, this did not account for the case where the checkout is using the
"share" extension in Hg where the checkout "shares" the `.hg/store` folder with
another local checkout.

This change updates `HgService` to check for `.hg/sharedpath`, indicating the
use of the share extension. If found, it uses the contents of the file to
identify the location of the local checkout whose `.hg/store` folder should be
monitored.

Reviewed By: tjfryan

Differential Revision: D7659226

fbshipit-source-id: 7f921ab78b850feaa70a5d51313d1ad296e16bec