when local storage is corrupt and we fail to fetch the server connection config, fallback to trying to get it from other windows if they exist
Summary:
This seems to fix the corrupt store + multiple windows issue. The problem is that when we open a new window and the store is corrupt, we fallback to reauthenticating which kills and restarts the nuclide server. This, in turn, destroys the remote connection for other windows.

The fix is to check with other windows to see if they have a valid server config for the host before giving up.

Reviewed By: hansonw

Differential Revision: D7687510

fbshipit-source-id: e8f2c0273b7be0dbbb2b7deb2954fca803fc008e