Major refactoring of the nuclide-remote-atom-rpc package.
Summary:
This cleans up the `nuclide-remote-atom-rpc` in a number of ways.

Previously, `CommandServer` was an unofficial singleton that had a ton of
`static` methods and members. Now it is explicitly a singleton (the new
comments direct the user to use `command-server-singleton.js`) and all of the
`static` members have been removed.

Previously, the `AtomCommands` interface was overloaded as both the interface
that an Atom window should implement as well as the interface for tools that
wanted to communicate with one or more Atom windows. With the introduction of
the `getClientConnections()` API in D7598900, using this interface for both
cases started to show its weakness. This revision formally separates the
interfaces into `AtomCommands` and `MultiConnectionAtomCommands`.

Further, `CommandServer.js` was broken up so that `RoutingAtomCommands`
(formerly `ServiceAtomCommands`) and `CommandServerConnection` were moved into
their own files.

While here, logic that was duplicated across `bin/nuclide-connections-node.js`
and `bin/main.js` were consolidated in `bin/CommandClient.js`.

Going forward, this should make it easier to create new tools that talk to Atom
clients now that new client capabilities can be added to `AtomCommands`
independently of changes to the `MultiConnectionAtomCommands` API that exposes
those capabilities to the tools.

Reviewed By: hansonw

Differential Revision: D7691934

fbshipit-source-id: 5aa7ef1626a83a32539ba01975d76423b225787d