Move commons node cache to nuclide commons SimpleCache
Summary: DevicePoller depends on cache/SimpleCache and I am planning on moving Android debugging to Atom IDE. To do that I need to move DevicePoller to modules which means I need to move commons-node/cache.js to nuclide-commons/SimpleCache.js. I added a TODO to that file to eventually merge it with nuclide-commons/cache.js but it's not important to do that right away.

Reviewed By: hansonw, a20012251

Differential Revision: D7736559

fbshipit-source-id: 246dddb562f8f236f5390ee6727c8c2e95a8e977