Have client specify whether to prefer custom file search rather than the server.
Summary:
A Nuclide server can support multiple clients, so rather than the server
making the decision regarding custom file search, let the client do it.

To make this work, the cache key of `DirectorySearchConfig` objects is now a
combination of the `preferCustomSearch` boolean provided by the client and the
directory to be searched.

I verified that all of the `isGkEnabled()` calls introduced by this revision
occur in Atom packages, so they are guaranteed to be performed on the client.
Note that the synchronous `isGkEnabled()` API is used rather than the async
`passesGK()` API because if we get the wrong value of this GK because not all of
the GKs have been loaded yet, it is not a huge deal.

Reviewed By: stepanhruda

Differential Revision: D7769179

fbshipit-source-id: 5d4ee00befeb628489c92df4caf8e42506c82aad