Fix a regression: remote `atom` command should fail gracefully if no connections.
Summary:
This is a regression I accidentally introduced in D7691934. The issue is that
I forgot to await a Promise such that this expression is guaranteed to
evaluate to `false`:

```
if (commands.getConnectionCount() === 0 && rejectIfZeroConnections) {
```

Because `commands.getConnectionCount()` returns a `Promise`, it will never be
equal to `0`. The fix is to add an `await`:

```
if ((await commands.getConnectionCount()) === 0 && rejectIfZeroConnections) {
```

Ideally this would have been caught by Flow.

Reviewed By: semmypurewal

Differential Revision: D7772751

fbshipit-source-id: 3e7162664ca323fec3a7f93492f98e45b40b82d3