Add 'observeHeadRevision' to HgRepositoryClient
Summary:
This seems like something we've been wanting, and it made sense to add it to HgRepositoryClient.

Should I add an argument to `distinctUntilChanged` that compares that hash instead of the object reference?

Unsure if it should be Observable<?RevisionInfo>, but that should only apply in cases where Revisions haven't yet loaded, or maybe empty repos.

Reviewed By: evangrayk

Differential Revision: D7799116

fbshipit-source-id: dbc49b10329ea72ad7a0a8dd76e800be30113633