atom-notify CLI for sending notifications from machine running Nuclide server to the client.
Summary:
This makes it possibly to run something like:

```
atom-notify hello world!
```

on your server, which will then take the arguments, join them on `' '`,
and display them as a message in an Atom notification. Although
Atom notifications have many options, the only ones that we
expose from the CLI for now are:

* `--success`
* `--info`
* `--warning`
* `--error`

which affects the UI treatment of the notification in Atom.

Reviewed By: hansonw

Differential Revision: D7719532

fbshipit-source-id: 1d95a667d7447cddc1b6c8878560ed2e6d355cec