Remove redundant feature group state
Summary:
Here we had some user-provided data that was then formatted into a more reasonable type. We kept both of those values, even though we only used the formatted one, and the formatting was implemented as a method that accessed the unformatted one. This is much clearer as explicit parameters.

We also exposed the formatted value even though it doesn't appear to be used anywhere so I removed that.

Reviewed By: wanderley

Differential Revision: D7801095

fbshipit-source-id: c9519e4fb6172569de792a06a0c8a6d0313ded17