Remove feature where F401 lints highlight the import's range
Summary:
There was this feature implemented a long time ago where F401 lint warnings (e.g. "F401 'datetime' imported but unused") would try to highlight the exact token that's imported (so in this case we'd underline 'datetime' rather than the whole line.)

It turns out that this isn't working anymore with the newer Python grammars, and it also causes performance issues because it ends up scanning the entire file. This could be fixed up a little bit but I think it's easier if we just get rid of it.

Reviewed By: pelmers

Differential Revision: D7835614

fbshipit-source-id: b2ae6c660c64c4e182ac07636ca6298a07b38666