immutable-snapshot: Delta snapshots of Immutable.JS objects
Summary:
`immutable-snapshot` provides custom extensions of Immutable.JS collections that enable the creation of serializable "delta snapshots" of a given Immutable object. These snapshots can then be sent over the wire and reconstructed with `ImmutableSnapshotReader`.

Delta snapshots have the ability to re-use objects in previous snapshots to avoid sending over objects that are already known. `immutable-snapshot` patches methods in the Immutable.JS implementation to log mutations to previously snapshotted objects, and can record the mutations rather than the entire object contents.

The primary motivation behind this kind of snapshot feature is to enable fast synchronization of large Immutable object trees (e.g. a Redux store) with a view (e.g. a React component) across process boundaries.

One example is the Nuclide console, which is essentially one big `Immutable.List` of console records. By using `immutable-snapshot` we have the ability to have a low-level stream of appends/edits/deletions based on the Redux state. while ensuring that all the logic occurs in a separate process.

This diff only implements support for `Immutable.List`. The following diff contains support for Map, OrderedMap, and Record.

I left a ton of comments in the code so please read those for the hairy implementation details.

Some design decisions that I made here (that are up for discussion):

- I decided on using a global mutations record after talking to matthewwithanm :)
  - While not as fine grained as tying each collection to a local store, this has the benefit of not needing to pass around a store everywhere.
- Monkey-patching via whitelist vs blacklist
  - I originally had it so that methods were patched via whitelist rather than blacklist, but this seems more robust. I haven't run the Immutable.JS test suite to see if the patching is breaking anything (but I should).
  - See https://github.com/kitten/extendable-immutable/tree/master/src for some related work.

Reviewed By: matthewwithanm

Differential Revision: D7721585

fbshipit-source-id: 9890d4a79d0375af547e9a53d3ffca98749151ce