Add Flow typings for vscode-jsonrpc
Summary:
I didn't realize this but the type import in `nuclide-commons-atom/ExperimentalMessageRouter` crosses back into `pkg`, which is no good :(

`vscode-jsonrpc` should have a real Flow library definition, so I'm adding it to `flow-libs` for now. There might be a few rough edges here so I'd like to battle-test it a bit before upstreaming it into flow-typed.

This exposed a pretty legit flaw in how we're using `createMessageConnection`: the logger argument is *not* optional!! Other than that, all the other changes are just corrections/improvements to make Flow happy.

Reviewed By: wanderley, arxanas

Differential Revision: D7852540

fbshipit-source-id: 57e7faf362adf9b8f538ca30fd0f8e4b7d71a67a