Java debugger: Fix bug preventing remote build + local debug
Summary: The debugger currently just hangs after creating the tunnel because this promise went unresolved. Fixing bug where if starting debugging fails (or is cancelled by the user), we leak the subscriptions disposable - which also leaks cleanupSubject, if this happens the user needs to restart Atom to be able to launch the Java debugger again.

Reviewed By: Goom11

Differential Revision: D7851456

fbshipit-source-id: c8aeecdcbbb37574e0f9424d035d1fe360ebccd6