configuration file
Summary:
We need the ability to be able to start common debugging scenarios without a ton of flags, and we also need a place to put configuration that's Facebook specific without (hopefully) putting it into the public code in try/require/catch blocks.

This diff adds a config file in a dot directory for that purpose. The config file can be pushed out via Chef and contains presets that alias to sets of command line arguments. As an added bonus, if fbdbg is run as a symlink named something else, then the symlink name will be used as the preset. So link hphpd to fbdbg, fbdbg will be run with '--preset hphpd' and will pick up all the settings to attach to hphpd with the correct target.

Reviewed By: ebluestein

Differential Revision: D7852817

fbshipit-source-id: a6e72d3ba23b742ec0245ebe0a5ae0ed04b5c3d2