Nuclide debugger: Put focus back in console prompt after selecting stop line
Summary: Currently when the debugger stops, it puts the focus in a text editor on the current stop line. We want to highlight the current line, but it's nice to leave the cursor in the console prompt so the user can type into the REPL without having to touch the mouse.

Differential Revision: D7882445

fbshipit-source-id: 06796e55a7c3f373fc0f6fb21fa2997db4412a4e