Make the user selected deployment target sticky between devices changes
Summary:
The problem I'm trying to solve here is when a device gets disconnected and then connected again later. When this happens and the selected deployment target was this device we don't re-select it once the device is connected again.

This is particularly worrying in tunnel based scenarios since sometimes the tunnel can be flaky and the device disconnects & connects due to this.

I solve this by adding a new state named `userSelectedDeploymentTarget` that the actual user selected deployment target (as opposed to one selected by the system).

Reviewed By: shunfu

Differential Revision: D7891892

fbshipit-source-id: c545804b22a4384040926fb58f7ded2096a8d9d1