handle ENOENT error when reusing an existing server connection
Summary:
We were swallowing the ENOENT error and returning null from RemoteConnection.reconnect when reusing an existing connection. This forced us to reattempt an ssh handshake, which restarted the server making the current certificates invalid.

This diff fixes this by capturing rethrowing the original ENOENT, and setting up an onError handler for the observable in the connection dialog.

Reviewed By: hansonw

Differential Revision: D7882443

fbshipit-source-id: 4e95bf64a3b860980c9dd28a72bcbb2c2681d490