Nuclide Java Debugger: Fix target crash and immediate VM disconnect when attaching to an Android target with lots of classes
Summary:
Requesting a list of all loaded classes was done so that breakpoints set on classes that were already loaded when the debugger attached would still work. Such classes are never going to generate a ClassPreparedEvent, because they're already loaded by the JVM.

Unforutnately, calling this crashes ART/Dalvik in cases where the target app has too many classes - the VM side tries to keep a weak reference for each ReferenceType mirror created and returned over JDWP, and this quickly exhausts the weak reference table in the GC, and the target sends itself a SIGABRT, from which we cannot recover.

This change makes it so we don't need to grab the full list - we parse the file with the breakpoint in it to get the class name for the breakpoint, and ask for only the classes that we actually need.

NOTE: This requires the file to exist and be parsable by the Java debugger. In the case of source+line breakpoints, we expected this already anyway.

Reviewed By: drmathochist

Differential Revision: D7872423

fbshipit-source-id: 135cab33b46caa73bebb06f3e4f482fb57b2534e