Make scribe appender resilient to circular objects in logged event
Summary:
The scribe appender uses JSON.stringify to convert object data to JSON before sending it. JSON.stringify will throw an exception if there are circular dependencies in the object (and I've seen this happen in logs).

Bring in a package that handles breaking the cycles so that if they happen, we don't lose the whole event.

JSON.stringify is everywhere in our code, and there may be other places that are problematic.

Reviewed By: hansonw

Differential Revision: D7889695

fbshipit-source-id: eeead54db5bb4df6eda11a17d573ff6994669a1d