Don't create dangling ServerConnection objects
Summary:
In `RemoteConnection.findOrCreateFromConnection`, we saw that sometimes `resolveRealPath` was failing. semmypurewal made sure that ENOENT was handled in a cleaner way, but this still potentially creates dangling `ServerConnection` objects that aren't being used.

We can catch all errors in `findOrCreateFromConnection` and close the server connection if an error occurred and no other `RemoteConnection` objects are using the connection.

I also went ahead and merged `findOrCreate` and `findOrCreateFromConnection` because it wasn't being used from anywhere else. (The separation doesn't add a lot and creates the illusion that there are other callsites.)

Reviewed By: semmypurewal

Differential Revision: D7900187

fbshipit-source-id: ddeef9ed3c4da0c1297a585f8eadcc27e2a63fdb