Fix indentation in outline view
Summary:
Showing outline view on this test-spec.js file:

```js
function topLevel1() {}

describe('Should be unindented', () => {
  it('A should be indented 1 level', () => {
  });

  describe('B should be indented 1 level', () => {
    it('BA should be indented 2 levels', () => {
    });
  });
  it('C should be indented 1 level', () => {
  });

});

function topLevel2() {}
```

Before yields:

{F125213807}

Note that the A and C nodes are incorrectly indented; they should be indented to match up with the B node.

After yields:

{F125213851}

Note that this fixes the indenting of the 'A' and 'C' entries in the outline view.

Reviewed By: matthewwithanm

Differential Revision: D7900313

fbshipit-source-id: 1e6b256ea6a74a40e15b4b9a8b8cd8e4773c6d1e