Stop existing tunnels when a new create tunnel is issued for the same port
Summary:
When a tunnel is created from the server to the laptop, nuclide server is listening on the from port.
If nuclide client reloads, it loses access to which tunnels are created on the server and when it tries to recreate them it will fail because the port is already bound to.

To get around this issue I'm keeping a map of all the tunnels created and will dispose them if the service receives a request to create them again (effectively working as a re-create).

I thought about returning the observable that is already stored in the case of the tunnel already existing, but I didn't want to make so many assumptions, specially because the expectation from these functions is that they're returning a connectable observable that is not connected to yet (which wouldn't be the case anymore if re-using the same observable).

Reviewed By: shunfu

Differential Revision: D7891580

fbshipit-source-id: 7019e75e21b1318e58a7da4f63e9a3da32d6c717